+++ src/main/java/util/Longitude.java
@@ -48,8 +48,8 @@ public class Longitude extends Coordinate {
 
     // JavaDoc made in Coordinate
     public double getDecimalCoordinate() {
-        double decimalCoordinate; 
-        decimalCoordinate = this.minutes + this.seconds/60;
+        double decimalCoordinate;
+        decimalCoordinate = this.minutes + this.seconds/(double)60;
         decimalCoordinate = this.degree + decimalCoordinate/60;
 
         // If the direction is West, then it turns negative
